requirements.txt
awscli==1.11.182
awsebcli==3.12.1
beautifulsoup4==4.6.0
blessed==1.14.2
boto==2.48.0
boto3==1.4.7
botocore==1.7.40
cement==2.8.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
DateTime==4.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.9
django-allauth==0.34.0
django-contrib-comments==1.8.0
django-crispy-forms==1.7.0
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-storages==1.6.5
django-taggit==0.22.1
djangorestframework==3.7.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
markdown2==2.3.4
MarkupSafe==1.0
nba-py==0.1.1a2
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.0
pathspec==0.5.0
Pillow==3.1.0
praw==5.2.0
prawcore==0.12.0
psycopg2==2.7.3.2
py-goldsberry==1.0.1
pyasn1==0.3.7
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.11
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
update-checker==0.16
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.12.2
zope.interface==4.4.3
